requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
gunicorn
django-heroku
pytest
pytest-django
dj-database-url
model_bakery